Clean the clippy gate, and put the default seed in one place
clippy found a stale assertion that had become vacuous: PLUS4_BASE moved to
zero when the charset budget was worked out, and screen_code(g) >= 0 is
always true.  Replaced with the property that actually matters - the Plus/4
screen-code mapping is the identity - which is what the target's inner loop
depends on and what should fail if the base ever moves again.

The rest is tidying, except one real defect.  The Makefile's SEED did not
match the seed the binaries default to, so 'make prg' baked a different city
into the Plus/4 build than the terminal rendered.  Nothing failed; the two
targets just quietly showed different places.  There is now one constant,
ascitty_core::DEFAULT_SEED, and the Makefile carries it written out for make
with a comment saying so.

Also: world.rs now has a Rect rather than passing four loose corners, which
is what clippy's too-many-arguments was really pointing at.

+++ crates/ascitty-core/src/raycast.rs
@@ -7,8 +7,8 @@
 //! enough for a skyline, because the thing that makes a city look like a
 //! city is a *tall* building visible over the top of a *near* one.  So the
 //! walk does not stop at the first hit.  It keeps going, front to back,
-//! carrying one number - the topmost screen row anything has claimed so far
-//! - and each further building may only draw above that line.  When the line
+//! carrying one number (the topmost screen row anything has claimed so far),
+//! and each further building may only draw above that line.  When the line
 //! reaches the top of the screen the column is finished and the walk stops.
 //!
 //! That is the Comanche voxel-space idea rather than the Wolfenstein one,
@@ -457,7 +457,7 @@ mod tests {
             for deg in (0..360).step_by(11) {
                 let mut cam = Camera::spawn(&city, 20, 70);
                 cam.yaw = crate::trig::from_degrees(deg as f64);
-                cam.pitch = (deg % 21) as i32 - 10;
+                cam.pitch = (deg % 21) - 10;
                 render(&city, &cam, &atmos, &mut f);
             }
         }
